backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
pydantic==2.11.1
pydantic-settings==2.8.1
python-dotenv==1.1.0
httpx==0.28.1
google-genai==1.14.0
sse-starlette==2.2.1
